clicp2icns 1.0d1
Derivative of clip2cicn that could be used to create icns resources
from the clipboard. A 64x32 clipboard image would be turned into a
32x32 32-bit icon (with an 8-bit mask) and saved as an resource
with an ID of the user's choosing.

icns resources were introduced in Mac OS 8.5, and ResEdit was never
 updated to handle them. While icons could be set via the Finder
"Get Info" window, that did not provide much control over the mask.