Badger 1.0d1
Badger started out at the suggestion/request of Arlo Rose (who also
at least partly the motivation for the creation of clip2icns). He was
in the process of creating 32-bit versions of the system folder icons
of the 5 built-in Kaleidoscope schemes, which was going to involve
hand-creating 48 * 5 = 240 icons. This being a time before
https://xkcd.com/1319/, he asked if I could create a tool that could
combine 48 folder "badges" with a different base icon for each scheme,
thus removing a lot of the tedium.

The initial version combines a base folder icon (ID -3999) with all
of the other icns resources found in a badge set file. There's no
mechanism for positioning the badgees, and icon manipulation is done
with the Icon Services API introduced in Mac OS 8.5, which cannot
generate old-style icons.