clip2icns 1.1
Handle files without resource forks (create one)

Name generation can be turned off in the preferences dialog.

Adds easter egg in about box.

Now distributed under the Mscape Software label (instead of via
GUI Central)

Old style icons can be generated too, instead of just picking up
existing ones (but chances are that they will look horrible).

The source archive was labeled as 1.1 but it does not appear to
contain some of the features described in the Read Me (e.g. being
able to type in an ID and have the appropriate icon be selected
in the menu) so it may actually be an earlier snapshot.