clip2icns 1.1.1
Don't hardcode the size of the image when placing it in the clipboard,
it may not always be 4662 bytes (see also
http://blog.persistent.info/1998/12/swatch-y-control.html)

Add a Photoshop action for converting a layer (and its alpha
channel) into a 64x32 image that's ready to be pasted.

Fix some smaller bugs (that were all in icnsClass and thus were
not part of the source snapshot).

Also includes some of the code changes that were missing
from the 1.1 source snapshot:
- Website and email address in the about box are clickable
- Select icon from popup that matches a manually-typed in ID