Glypher 1.0d8
Check for a registered copy of clip2icns, otherwise a watermark
(a stylized G) is added to generated icons.

Allow glyphs to be dragged around with the mouse or the arrow keys
(option moves the small glyph instead of the large one).

When holding down shift while clicking the position buttons, move
glyphs by 10 pixels instead of 1.

Add email and website links to the about box.

Add more options for which file to insert merged glyphs into (new
file, or an existing Kaleidoscope Scheme)