Icon Mangler 1.0d2
Icon editor project that I started with the dual goals of serving as
my IB Computer Science project and something to sell as shareware, as
a full-featured follow-on to clip2icns.

This version is one of the earliest that I distributed to testers on a
Hotline server that I was on. It has the following features:
- Loading and saving of 48x48, 32x32 and 16x16 icons at 32-, 8-, 4- and
  1-bit depths
- Basic "browser" for going through files with multiple icons (e.g.
  Kaleidoscope schemes) and choosing which one to edit
- Basic drawing commands: pencil, eraser, pain bucket/flood fill, eye dropper,
  line, rectangle, selections (rectangular, lasso, magic wand)
- Rotation and flip transforms
- Nearest neighbor zoom to make it easier to work on small icons
- Appearance Manager support for UI controls

See the journal entries from http://blog.persistent.info/search/label/Iconographer?updated-max=1999-01-20T08:00:00-08:00&by-date=true
for more details