Icon Mangler 1.0d4
Continuation of icon editor work. In preparation for submission to the IB the
code was commented (to a fault) and lots of error checking was added (mostly
for low memory conditions).

See the journal entries from http://blog.persistent.info/search/label/Iconographer?updated-max=1999-04-19T08:00:00-08:00&by-date=true
for more details